refactor(vm): isolate module loading
